Move Process to `base.py`
This is required to avoid circular dependencies when we inherit
Synapse from Process in the next commit, since `processes.py` uses
things from `synapses.py`.
