+++ nengo/ensemble.py
@@ -1,11 +1,10 @@
 import weakref
 
-from nengo.base import NengoObject, ObjView
+from nengo.base import NengoObject, ObjView, ProcessParam
 from nengo.dists import DistOrArrayParam, Uniform, UniformHypersphere
 from nengo.exceptions import ReadonlyError
 from nengo.neurons import LIF, NeuronTypeParam, Direct
 from nengo.params import Default, IntParam, NumberParam
-from nengo.processes import ProcessParam
 
 
 class Ensemble(NengoObject):
